requirements.txt
streamlit
pandas
underthesea
gensim
seaborn
matplotlib.pyplot
sklearn
pickle
PIL
base64